+++ extensions/promptlab/src/components/Commands/CommandListDetail.tsx
@@ -94,8 +94,8 @@ ${
         command.minNumFiles == "0"
           ? "N/A"
           : command.acceptedFileExtensions?.length && command.acceptedFileExtensions !== "None"
-          ? command.acceptedFileExtensions
-          : "Any"
+            ? command.acceptedFileExtensions
+            : "Any"
       } |
 | Creativity | ${command.temperature == undefined || command.temperature == "" ? "1.0" : command.temperature} |
 | Use File Metadata? | ${isTrueStr(command.useMetadata) ? "Yes" : "No"} |
@@ -104,6 +104,7 @@ ${
 | Use Audio Transcription? | ${isTrueStr(command.useAudioDetails) ? "Yes" : "No"} |
 | Use Barcode Detection? | ${isTrueStr(command.useBarcodeDetection) ? "Yes" : "No"} |
 | Use Face Detection? | ${isTrueStr(command.useFaceDetection) ? "Yes" : "No"} |
+| Use Horizon Detection? | ${isTrueStr(command.useHorizonDetection) ? "Yes" : "No"} |
 | Use Rectangle Detection? | ${isTrueStr(command.useRectangleDetection) ? "Yes" : "No"} |
 | Use Saliency Analysis? | ${isTrueStr(command.useSaliencyAnalysis) ? "Yes" : "No"} |
 ${isCommand(command) ? `| Model | ${command.model || "Not Specified"} |` : ``}
@@ -119,7 +120,7 @@ ${(isCommand(command) ? command.setupConfig : JSON.parse(command.setupConfig)).f
     (field: NumberConfigField | BooleanConfigField | StringConfigField) =>
       `| ${field.name} | ${
         field.description == undefined || field.description.trim().length == 0 ? "None" : field.description
-      } | ${field.value == undefined || field.value.toString().trim().length == 0 ? "None" : field.value} |`
+      } | ${field.value == undefined || field.value.toString().trim().length == 0 ? "None" : field.value} |`,
   )
   .join("\n")}
 `