+++ extensions/reclaim-ai/src/hooks/useTask.tsx
@@ -1,3 +1,7 @@
+import { getPreferenceValues } from "@raycast/api";
+import { useFetch } from "@raycast/utils";
+import { useMemo } from "react";
+import { NativePreferences } from "../types/preferences";
 import { Task } from "../types/task";
 import { axiosPromiseData } from "../utils/axiosPromise";
 import reclaimApi from "./useApi";
@@ -5,6 +9,23 @@ import { ApiResponseTasks, CreateTaskProps } from "./useTask.types";
 
 const useTask = () => {
   const { fetcher } = reclaimApi();
+
+  const { apiUrl, apiToken } = getPreferenceValues<NativePreferences>();
+
+  const headers = useMemo(
+    () => ({
+      Authorization: `Bearer ${apiToken}`,
+      "Content-Type": "application/json",
+      Accept: "application/json",
+    }),
+    [apiToken]
+  );
+
+  const useFetchTasks = () =>
+    useFetch<ApiResponseTasks>(`${apiUrl}/tasks`, {
+      headers,
+      keepPreviousData: true,
+    });
 
   const createTask = async (task: CreateTaskProps) => {
     try {
@@ -18,7 +39,8 @@ const useTask = () => {
         minChunkSize: task.durationMin,
         maxChunkSize: task.durationMax,
         notes: task.notes,
-        alwaysPrivate: true,
+        priority: task.priority,
+        onDeck: task.onDeck,
       };
 
       const [createdTask, error] = await axiosPromiseData<Task>(
@@ -55,7 +77,7 @@ const useTask = () => {
     }
   };
 
-  const fetchTasks = async () => {
+  const getAllTasks = async () => {
     try {
       const [tasks, error] = await axiosPromiseData<ApiResponseTasks>(fetcher("/tasks"));
       if (!tasks && error) throw error;
@@ -65,11 +87,73 @@ const useTask = () => {
     }
   };
 
+  // Add time
+  const addTime = async (task: Task, time: number) => {
+    try {
+      const [updatedTime, error] = await axiosPromiseData(
+        fetcher(`/planner/add-time/task/${task.id}?minutes=${time}`, { method: "POST", responseType: "json" })
+      );
+      if (!updatedTime || error) throw error;
+      return updatedTime;
+    } catch (error) {
+      console.error("Error while adding time", error);
+    }
+  };
+
+  // Update task
+  const updateTask = async (task: Task) => {
+    try {
+      const [updatedTask, error] = await axiosPromiseData(
+        fetcher(`/tasks/${task.id}`, {
+          method: "PUT",
+          responseType: "json",
+          data: task,
+        })
+      );
+
+      if (!updatedTask || error) throw error;
+      return updatedTask;
+    } catch (error) {
+      console.error("Error while updating task", error);
+    }
+  };
+
+  // Set task to done
+  const doneTask = async (task: Task) => {
+    try {
+      const [updatedStatus, error] = await axiosPromiseData(
+        fetcher(`/planner/done/task/${task.id}`, { method: "POST", responseType: "json" })
+      );
+      if (!updatedStatus || error) throw error;
+      return updatedStatus;
+    } catch (error) {
+      console.error("Error while updating task", error);
+    }
+  };
+
+  // Set task to incomplete
+  const incompleteTask = async (task: Task) => {
+    try {
+      const [updatedStatus, error] = await axiosPromiseData(
+        fetcher(`/planner/unarchive/task/${task.id}`, { method: "POST", responseType: "json" })
+      );
+      if (!updatedStatus || error) throw error;
+      return updatedStatus;
+    } catch (error) {
+      console.error("Error while updating task", error);
+    }
+  };
+
   return {
+    useFetchTasks,
     createTask,
-    fetchTasks,
     handleStartTask,
     handleStopTask,
+    getAllTasks,
+    addTime,
+    updateTask,
+    doneTask,
+    incompleteTask,
   };
 };